requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==5.1.4
django-ckeditor==6.7.2
django-jazzmin==3.0.1
django-js-asset==3.0.1
django-paypal==2.1
django-taggit==6.1.0
idna==3.10
mysqlclient==2.2.7
pillow==11.1.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
shortuuid==1.0.13
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
